test(agents): align runner suites with main
